% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carte_gini_spe.R
\name{carte_gini_spe}
\alias{carte_gini_spe}
\title{Réaliser la carte de l'indice de Gini de spécificité globale}
\usage{
carte_gini_spe(bdd, carte = NA, codgeo1 = NA, codgeo2 = NA,
  methode = "fisher-jenks", contour = TRUE,
  titre = "Indice de specialisation de Gini", n = 5, cols = carto.pal(pal1
  = "purple.pal", n1 = n))
}
\arguments{
\item{bdd}{data.frame. La base de données avec en première colonne le code géographique
et les effectifs des différentes catégories dans les colonnes suivantes.}

\item{carte}{spdf. Fond de carte à la même échelle que la base de données. Par defaut NA.
Si non renseigné aucune carte n'est affichée.}

\item{codgeo1}{character. Nom de la variable correspondant au code géographique dans le fond de carte.
Par défaut vaut NA et la première colonne est considérée comme code géographique.}

\item{codgeo2}{character. Nom de la variable correspondant au code géographique dans la base de données.
Par défaut vaut NA et la première colonne est considérée comme code géographique.}

\item{methode}{string. Méthode de discrétisation de la carte. Voir \link[cartography]{choroLayer}
pour les différentes méthodes possibles. Par défaut la méthode est "fisher-jenks".}

\item{contour}{boolean. Indique si le contour du fond plan doit être affiché sur la carte.
Par défaut vaut TRUE.}

\item{titre}{character. Titre de la carte. Par défaut il est "Indice de specialisation de Gini".}

\item{n}{int. Nombre de classes à faire dans la carte. Par défaut : 5.}

\item{cols}{character. Les couleurs à utiliser pour la carte. Voir \link[cartography]{carto.pal}
pour les palettes de couleur. Par défaut : carto.pal(pal1 = "purple.pal", n1 = 5)}
}
\value{
Affiche la carte de l'indice de Gini de spécificité globale.
}
\description{
Cette fonction réalise la carte de l'indice de Gini de spécificité globale.
}
\examples{
# mettre un exemple


}
