% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carte_indice_specif.R
\name{carte_indice_specif}
\alias{carte_indice_specif}
\title{Réaliser la carte de l'indice de spécificité de Kubrak pour une catégorie}
\usage{
carte_indice_specif(bdd, numCol, diff = FALSE, carte, codgeo1 = NA,
  codgeo2 = NA, methode = "fisher-jenks", contour = TRUE, titre = NA,
  n = 5, cols = carto.pal(pal1 = "purple.pal", n1 = n))
}
\arguments{
\item{bdd}{data.frame. La base de données avec en première colonne le code géographique
et les effectifs des différentes catégories dans les colonnes suivantes.}

\item{numCol}{int. Indique le numéro de colonne de la catégorie à considérer (en comptant 
la 1ere colone de la zone géographique).}

\item{diff}{boolean. Indique si l'indice de spécificité doit être calculé par différence
(diff=TRUE) ou par ratio (diff=FALSE). C'est-à-dire si on fait la différence par
rapport au reste du territoire de la part d'une zone dans une catégorie ou le ratio.
Par défaut il vaut FALSE.}

\item{carte}{spdf. Fond de carte à la même échelle que la base de données. 
Si non renseigne aucune carte n'est affichee.}

\item{codgeo1}{character. Nom de la variable correspondant au code géographique dans le fond de carte.
Par défaut vaut NA et la première colonne est considérée comme code géographique.}

\item{codgeo2}{character. Nom de la variable correspondant au code géographique dans la base de données.
Par défaut vaut NA et la première colonne est considérée comme code géographique.}

\item{methode}{string. Méthode de discrétisation de la carte. Voir \link[cartography]{choroLayer}
pour les différentes méthodes possibles. Par défaut la méthode est "fisher-jenks".}

\item{contour}{boolean. Indique si le contour du fond plan doit être affiché sur la carte.
Par défaut vaut TRUE.}

\item{titre}{character. Titre de la carte. Par défaut, il est "Indice de specicite de " suivi du nom de
la catégorie considérée.}

\item{n}{int. Nombre de classes à faire dans la carte. Par défaut : 5.}

\item{cols}{character. Les couleurs à utiliser pour la carte. Voir \link[cartography]{carto.pal}
pour les palettes de couleur. Par défaut : carto.pal(pal1 = "purple.pal", n1 = 5)}
}
\value{
Affiche la carte de l'indice de spécificité de Kubrak calculé pour une catégorie.
}
\description{
Cette fonction réalise la carte de l'indice de spécificté de Kubrak pour une catégorie.
L'indice de spécificté de Kubrak est calculé pour chaque zone et la catégorie demandée.
r(ik) = X(ik)/X(k)  / ( (X(k) - X(ik)) / (X-X(i)) ) (par ratio)
r(ik) = X(ik)/X(k)  - ( (X(k) - X(ik)) / (X-X(i)) ) (par différence)
où i correspond à la zone et k à la catégorie
}
\examples{
# mettre un exemple

}
