% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini_spe.R
\name{gini_spe}
\alias{gini_spe}
\title{Calculer l'indice de Gini de spécificité globale}
\usage{
gini_spe(bdd)
}
\arguments{
\item{bdd}{data.frame. La base de données avec en première colonne le code géographique
et les effectifs des différentes catégories dans les colonnes suivantes.}
}
\value{
data.frame. Renvoit la base de données initiale avec une colonne supplémentaire contenant
l'indice de Gini de spécificité globale.
}
\description{
Cette fonction calcule l'indice de Gini de spécificité globale qui est une mesure de
l'inégalité d'une distribution. Il est compris entre 0 et 1.
}
\examples{
bdd <- data.frame(zones=c("zone1","zone2","zone3","zone4"),categorie1=c(22,14,7,55),
categorie2=c(32,17,12,9),categorie3=c(41,32,10,16))
gini_spe(bdd)

}
