% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indice_specif.R
\name{indice_specif}
\alias{indice_specif}
\title{Cette fonction calcule l'indice de spécificté de Kubrak}
\usage{
indice_specif(bdd, diff = FALSE)
}
\arguments{
\item{bdd}{data.frame. La base de données avec en première colonne le code géographique
et les effectifs des différentes catégories dans les colonnes suivantes.}

\item{diff}{boolean. Indique si l'indice de spécificité doit être calculé par différence
(diff=TRUE) ou par ratio (diff=FALSE). C'est-à-dire si on fait la différence par
rapport au reste du territoire de la part d'une zone dans une catégorie ou le ratio.
Par défaut il vaut FALSE.}
}
\value{
data.frame. La base de données avec autant de colonnes supplémentaires que de catégories
et contenant l'indice de spécificité de cette catégorie pour une zone géographique par rapport au
reste du territoire. Les nouvelles colonnes se nomment "indic_specif_" suivi du nom de la
variable initiale.
}
\description{
L'indice de spécificté de Kubrak est calculé pour chaque zone et chaque catégorie.
r(ik) = X(ik)/X(k)  / ( (X(k) - X(ik)) / (X-X(i)) ) (par ratio)
r(ik) = X(ik)/X(k)  - ( (X(k) - X(ik)) / (X-X(i)) ) (par différence)
}
\examples{
bdd <- data.frame(zones=c("zone1","zone2","zone3","zone4"),categorie1=c(22,14,7,55),
categorie2=c(32,17,12,9),categorie3=c(41,32,10,16))
indice_specif(bdd)
indice_specif(bdd,diff=TRUE)


}
