% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krugman.R
\name{krugman}
\alias{krugman}
\title{Cette fonction calcule le coefficient de specificte globale de Krugman}
\usage{
krugman(bdd, carte = NA, codgeo1 = "NA", codgeo2 = "NA",
  methode = "fisher-jenks", contour = TRUE, titre = "Indice de Krugman",
  n = 5, cols = carto.pal(pal1 = "purple.pal", n1 = n))
}
\arguments{
\item{bdd}{data.frame. La base de donnees avec en première colonne le code geographique
et les effectifs des differentes categories dans les colonnes suivantes.}

\item{carte}{spdf. Fond de carte a la même echelle que la base de donnees. Par defaut NA.
Si non renseigne aucune carte n'est affichee.}

\item{codgeo1}{character. Nom de la variable correspondant au code geographique dans le fond de carte}

\item{codgeo2}{character. Nom de la variable correspondant au code geographique dans la base de donnees}

\item{methode}{string. Methode de discretisation de la carte. Voir \link[cartography]{choroLayer}
pour les differentes methodes possibles. Par defaut la methode est "fisher-jenks".}

\item{contour}{boolean. Indique si le contour du fond plan doit être affiché sur la carte.
Par défaut vaut TRUE.}

\item{titre}{character. Titre de la carte. Par defaut il est "Indice de Krugman".}

\item{n}{int. Nombre de classes a faire dans la carte. Par defaut : 5.}

\item{cols}{character. Les couleurs a utiliser pour la carte. Voir \link[cartography]{carto.pal}
pour les palettes de couleur. Par defaut : carto.pal(pal1 = "purple.pal", n1 = 5)}
}
\value{
data.frame. Renvoit la base de donnees initiale avec une colonne supplementaire contenant le
coefficient de Krugman.
}
\description{
Le coefficient de Krugman est un indice de specificite globale d'une zone par rapport
au reste du territoire, il est calcule pour chaque zone. Ce coefficient est compris entre
0 et 1, s'il est egal a 1 le territoire est très specifique par rapport aux autres.
}
\examples{
bdd <- data.frame(zones=c("zone1","zone2","zone3","zone4"),categorie1=c(22,14,7,55),
categorie2=c(32,17,12,9),categorie3=c(41,32,10,16))
krugman(bdd)


}
