% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediane.R
\name{mediane}
\alias{mediane}
\title{Cette fonction calcule un coefficient de spécificté globale comparant chaque zone
à la médiane du reste du territoire}
\usage{
mediane(bdd, part = FALSE)
}
\arguments{
\item{bdd}{data.frame. La base de données avec en première colonne le code géographique
et les effectifs ou part des différentes catégories dans les colonnes suivantes.}

\item{part}{boolean.  Indique si les données sont remplies en nombre (part=FALSE) ou
en part (part=TRUE) : dans ce cas, la somme des valeurs d'une ligne vaut 1.
Par défaut il vaut FALSE.}
}
\value{
data.frame. Renvoit la base de données initiale avec une colonne supplémentaire contenant le
coefficient de spécificté globale comparant chaque zone à la médiane
}
\description{
Cette fonction calcule un coefficient de spécificté globale comparant chaque zone
à la médiane
}
\examples{
bdd <- data.frame(zones=c("zone1","zone2","zone3","zone4"),categorie1=c(22,14,7,55),
categorie2=c(32,17,12,9),categorie3=c(41,32,10,16))
mediane(bdd)
mediane(bdd,part=TRUE)

}
