% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificite.R
\docType{package}
\name{specificite-package}
\alias{specificite-package}
\alias{specificite}
\title{specificite permet de calculer des indices de spécificité d'une zone géographique}
\description{
Ce package a pour but de calculer différents indices permettant de connaitre la spécificité
d'une zone géographique par rapport au reste du territoire.
}
\details{
\tabular{ll}{
  Package: \tab specifite\cr
  Type: \tab Package\cr
  Version: \tab 1.0.0\cr
  Date: \tab 2018-07-03\cr
  License: \tab GPL-2\cr
  LazyLoad: \tab yes\cr
}
}
\examples{
# ici il faut mettre un exemple de l'utilisation des fonctions de votre package
}
\author{
Elodie Molitor
}
\keyword{package}
