% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krugman.R
\name{krugman}
\alias{krugman}
\title{Calculer le coefficient de spécificité globale de Krugman, réaliser la carte}
\usage{
krugman(bdd, carte = NA, codgeo1 = NA, codgeo2 = NA,
  methode = "fisher-jenks", contour = TRUE, titre = "Indice de Krugman",
  n = 5, cols = carto.pal(pal1 = "purple.pal", n1 = n))
}
\arguments{
\item{bdd}{data.frame. La base de données avec en première colonne le code géographique
et les effectifs des différentes catégories dans les colonnes suivantes.}

\item{carte}{spdf. Fond de carte à la même échelle que la base de données. Par défaut NA.
Si non renseigné aucune carte n'est affichée.}

\item{codgeo1}{character. Nom de la variable correspondant au code géographique dans le fond de carte.
Par défaut vaut NA et la première colonne est considérée comme code géographique.}

\item{codgeo2}{character. Nom de la variable correspondant au code géographique dans la base de données.
Par défaut vaut NA et la première colonne est considérée comme code géographique.}

\item{methode}{string. Méthode de discrétisation de la carte. Voir \link[cartography]{choroLayer}
pour les différentes méthodes possibles. Par défaut la méthode est "fisher-jenks".}

\item{contour}{boolean. Indique si le contour du fond plan doit être affiché sur la carte.
Par défaut vaut TRUE.}

\item{titre}{character. Titre de la carte. Par défaut il est "Indice de Krugman".}

\item{n}{int. Nombre de classes à faire dans la carte. Par défaut : 5.}

\item{cols}{character. Les couleurs à utiliser pour la carte. Voir \link[cartography]{carto.pal}
pour les palettes de couleur. Par défaut : carto.pal(pal1 = "purple.pal", n1 = 5)}
}
\value{
data.frame. Renvoit la base de données initiale avec une colonne supplémentaire contenant le
coefficient de Krugman. 
Affiche seulement la carte si le fond de plan est donné.
}
\description{
Cette fonction calcule le coefficient de spécificté globale de Krugman.
Le coefficient de Krugman est un indice de spécificité globale d'une zone par rapport
au reste du territoire, il est calculé pour chaque zone. Ce coefficient est compris entre
0 et 1, s'il est égal à 1 le territoire est très specifique par rapport aux autres.
Cette fonction peut également afficher la carte de l'indice de Krugman lorsque les variables
nécessaires sont données à la fonction (fond de carte à la bonne échelle au minimum).
}
\examples{
bdd <- data.frame(zones=c("zone1","zone2","zone3","zone4"),categorie1=c(22,14,7,55),
categorie2=c(32,17,12,9),categorie3=c(41,32,10,16))
krugman(bdd)


}
